+++ mobile/android/android-components/components/service/glean/src/main/java/mozilla/components/service/glean/scheduler/GleanLifecycleObserver.kt
@@ -22,7 +22,7 @@ internal class GleanLifecycleObserver : LifecycleObserver {
     fun onEnterBackground() {
         // We're going to background, so store how much time we spent
         // on foreground.
-        GleanBaseline.duration.stopAndSum()
+        GleanBaseline.duration.stopAndSum(this)
         Glean.handleBackgroundEvent()
     }
 
@@ -39,6 +39,6 @@ internal class GleanLifecycleObserver : LifecycleObserver {
         // Note that this is sending the length of the last foreground session
         // because it belongs to the baseline ping and that ping is sent every
         // time the app goes to background.
-        GleanBaseline.duration.start()
+        GleanBaseline.duration.start(this)
     }
 }