+++ mobile/android/android-components/components/service/glean/src/main/java/mozilla/components/service/glean/storages/TimespansStorageEngine.kt
@@ -6,17 +6,14 @@ package mozilla.components.service.glean.storages
 
 import android.annotation.SuppressLint
 import android.content.SharedPreferences
-import android.os.SystemClock
 import android.support.annotation.VisibleForTesting
-import mozilla.components.service.glean.error.ErrorRecording.ErrorType
-import mozilla.components.service.glean.error.ErrorRecording.recordError
 import mozilla.components.service.glean.private.CommonMetricData
 import mozilla.components.service.glean.private.TimeUnit
+import mozilla.components.service.glean.utils.getAdjustedTime
 
 import mozilla.components.support.base.log.logger.Logger
 import org.json.JSONArray
 import org.json.JSONObject
-import java.util.concurrent.TimeUnit as AndroidTimeUnit
 
 /**
  * This singleton handles the in-memory storage logic for timespans. It is meant to be used by
@@ -32,12 +29,6 @@ internal object TimespansStorageEngine : TimespansStorageEngineImplementation()
 internal open class TimespansStorageEngineImplementation(
     override val logger: Logger = Logger("glean/TimespansStorageEngine")
 ) : GenericStorageEngine<Long>() {
-
-    /**
-     * A map that stores the start times from the API consumers, not yet
-     * committed to any store (i.e. [start] was called but no [stopAndSum] yet).
-     */
-    private val uncommittedStartTimes = mutableMapOf<String, Long>()
 
     /**
      * An internal map to keep track of the desired time units for the recorded timespans.
@@ -103,98 +94,32 @@ internal open class TimespansStorageEngineImplementation(
     }
 
     /**
-     * Helper function used for getting the elapsed time, since the process
-     * started, using a monotonic clock.
-     * We need to have this as an helper so that we can override it in tests.
-     *
-     * @return the time, in nanoseconds, since the process started.
-     */
-    internal fun getElapsedNanos(): Long = SystemClock.elapsedRealtimeNanos()
-
-    /**
-     * Convenience method to get a time in a different, supported time unit.
-     *
-     * @param timeUnit the required time unit, one in [TimeUnit]
-     * @param elapsedNanos a time in nanoseconds
-     *
-     * @return the time in the desired time unit
-     */
-    private fun getAdjustedTime(timeUnit: TimeUnit, elapsedNanos: Long): Long {
-        return when (timeUnit) {
-            TimeUnit.Nanosecond -> elapsedNanos
-            TimeUnit.Microsecond -> AndroidTimeUnit.NANOSECONDS.toMicros(elapsedNanos)
-            TimeUnit.Millisecond -> AndroidTimeUnit.NANOSECONDS.toMillis(elapsedNanos)
-            TimeUnit.Second -> AndroidTimeUnit.NANOSECONDS.toSeconds(elapsedNanos)
-            TimeUnit.Minute -> AndroidTimeUnit.NANOSECONDS.toMinutes(elapsedNanos)
-            TimeUnit.Hour -> AndroidTimeUnit.NANOSECONDS.toHours(elapsedNanos)
-            TimeUnit.Day -> AndroidTimeUnit.NANOSECONDS.toDays(elapsedNanos)
-        }
-    }
-
-    /**
-     * Start tracking time for the provided metric. This records an error if it’s
-     * already tracking time (i.e. start was already called with no corresponding
-     * [stopAndSum]): in that case the original start time will be preserved.
-     *
-     * @param metricData the metric information for the timespan
-     */
-    fun start(metricData: CommonMetricData) {
-        val timespanName = metricData.identifier
-
-        if (timespanName in uncommittedStartTimes) {
-            recordError(
-                metricData,
-                ErrorType.InvalidValue,
-                "Timespan already started",
-                logger
-            )
-            return
-        }
-
-        synchronized(this) {
-            uncommittedStartTimes[timespanName] = getElapsedNanos()
-        }
-    }
-
-    /**
-     * Stop tracking time for the provided metric. Add the elapsed time to the time currently
-     * stored in the metric. This will record an error if no [start] was called.
+     * Add the elapsed time to the time currently stored in the metric.
      *
      * @param metricData the metric information for the timespan
      * @param timeUnit the time unit we want the data in when snapshotting
+     * @param elapsedNanos the time to record, in nanoseconds
      */
     @Synchronized
-    fun stopAndSum(
+    fun sum(
         metricData: CommonMetricData,
-        timeUnit: TimeUnit
+        timeUnit: TimeUnit,
+        elapsedNanos: Long
     ) {
         // Look for the start time: if it's there, commit the timespan.
         val timespanName = metricData.identifier
-        uncommittedStartTimes.remove(timespanName)?.let { startTime ->
-            val elapsedNanos = getElapsedNanos() - startTime
 
-            // Store the time unit: we'll need it when snapshotting.
-            timeUnitsMap[timespanName] = timeUnit
+        // Store the time unit: we'll need it when snapshotting.
+        timeUnitsMap[timespanName] = timeUnit
 
-            // Use a custom combiner to sum the new timespan to the one already stored. We
-            // can't adjust the time unit before storing so that we still allow for values
-            // lower than the desired time unit to accumulate.
-            super.recordMetric(metricData, elapsedNanos, timeUnit) { currentValue, newValue ->
-                currentValue?.let {
-                    it + newValue
-                } ?: newValue
-            }
+        // Use a custom combiner to sum the new timespan to the one already stored. We
+        // can't adjust the time unit before storing so that we still allow for values
+        // lower than the desired time unit to accumulate.
+        super.recordMetric(metricData, elapsedNanos, timeUnit) { currentValue, newValue ->
+            currentValue?.let {
+                it + newValue
+            } ?: newValue
         }
-    }
-
-    /**
-     * Abort a previous [start] call. No error is recorded if no [start] was called.
-     *
-     * @param metricData the metric information for the timespan
-     */
-    @Synchronized
-    fun cancel(metricData: CommonMetricData) {
-        uncommittedStartTimes.remove(metricData.identifier)
     }
 
     /**