+++ mobile/android/android-components/components/service/glean/src/test/java/mozilla/components/service/glean/private/TimespanMetricTypeTest.kt
@@ -33,8 +33,8 @@ class TimespanMetricTypeTest {
         )
 
         // Record a timespan.
-        metric.start()
-        metric.stopAndSum()
+        metric.start(this)
+        metric.stopAndSum(this)
 
         // Check that data was properly recorded.
         assertTrue(metric.testHasValue())
@@ -54,11 +54,11 @@ class TimespanMetricTypeTest {
         )
 
         // Record a timespan.
-        metric.start()
-        metric.stopAndSum()
+        metric.start(this)
+        metric.stopAndSum(this)
 
         // Let's also call cancel() to make sure it's a no-op.
-        metric.cancel()
+        metric.cancel(this)
 
         // Check that data was not recorded.
         assertFalse("The API should not record a counter if metric is disabled",
@@ -78,13 +78,14 @@ class TimespanMetricTypeTest {
         )
 
         // Record a timespan.
-        metric.start()
-        metric.cancel()
-        metric.stopAndSum()
+        metric.start(this)
+        metric.cancel(this)
+        metric.stopAndSum(this)
 
         // Check that data was not recorded.
         assertFalse("The API should not record a counter if metric is cancelled",
             metric.testHasValue())
+        assertEquals(1, testGetNumRecordedErrors(metric, ErrorType.InvalidValue))
     }
 
     @Test(expected = NullPointerException::class)
@@ -113,8 +114,8 @@ class TimespanMetricTypeTest {
         )
 
         // Record a timespan.
-        metric.start()
-        metric.stopAndSum()
+        metric.start(this)
+        metric.stopAndSum(this)
 
         // Check that data was properly recorded in the second ping.
         assertTrue(metric.testHasValue("store2"))
@@ -134,9 +135,9 @@ class TimespanMetricTypeTest {
         )
 
         // Record a timespan.
-        metric.start()
-        metric.start()
-        metric.stopAndSum()
+        metric.start(this)
+        metric.start(this)
+        metric.stopAndSum(this)
 
         // Check that data was properly recorded in the second ping.
         assertTrue(metric.testHasValue("store2"))